+++ internal/setting/computed_minwinsvc.go
@@ -1,4 +1,4 @@
-// +build miniwinsvc
+// +build minwinsvc
 
 // Copyright 2015 The Gogs Authors. All rights reserved.
 // Use of this source code is governed by a MIT-style
@@ -11,5 +11,5 @@ import (
 )
 
 func init() {
-	SupportMiniWinService = true
+	supportWindowsService = true
 }