+++ internal/setting/setting.go
@@ -314,7 +314,6 @@ var (
 	// Other settings
 	ShowFooterBranding         bool
 	ShowFooterTemplateLoadTime bool
-	SupportMiniWinService      bool
 
 	// Global setting objects
 	Cfg          *ini.File